refactor(scoring): Konversi sistem penilaian dari integer ke decimal
- Change: Ubah bobot_pertanyaan di schema.sql dari INTEGER (20/10/60/10)
  menjadi FLOAT (0.2/0.1/0.6/0.1) untuk akurasi penilaian lebih baik
- Update: Ubah tipe data dataSkor di FormHasil dari Dictionary(Of String, Integer)
  menjadi Dictionary(Of String, Double)
- Refactor: Simplifikasi perhitungan faktorSkala di panelGrafik_Paint
- Fix: Update comparison logic untuk mendukung nilai skor desimal

Note: Masih ada bug di tampilkanData() saat highlight juara - CInt()
seharusnya CDbl() untuk menangani nilai desimal

+++ sistem-pakar/FormPertanyaan.vb
@@ -154,7 +154,7 @@
             End If
 
             'kirim array jawabanUser() ke fungsi cekRekomendasiTopik di ModuleInferensi
-            Dim hasilRekomendasi As Dictionary(Of String, Integer) = ModuleInferensi.cekRekomendasiTopik(jawabanUser, dtPertanyaan)
+            Dim hasilRekomendasi As Dictionary(Of String, Double) = ModuleInferensi.cekRekomendasiTopik(jawabanUser, dtPertanyaan)
 
             'tentukan juara 1 untuk disimpan ke database
             Dim juaraPertama As String = ""
